feat: entity 수정
