--- src/main/java/com/plantify/item/domain/entity/Status.java
@@ -1,5 +1,0 @@
-package com.plantify.item.domain.entity;
-
-public enum Status {
-    COMPLETED, CANCELED
-}
